Fix accrual balance add

+++ gophermart.go
@@ -42,9 +42,9 @@ type (
 	}
 
 	AccrualOrder struct {
-		Order   string `json:"order"`
-		Status  string `json:"status"`
-		Accrual int64  `json:"accrual"`
+		Order   string  `json:"order"`
+		Status  string  `json:"status"`
+		Accrual float64 `json:"accrual"`
 	}
 
 	AccrualClient interface {
